requirements.txt
glob2==0.7
json5==0.8.5
numba==0.45.1
numpy==1.16.5+mkl
opencv-python==4.1.2+contrib
pandas==0.25.1
tensorflow-gpu==2.0.0
